Support to restrict max physical relocation
(1)
Some older 32bit kernels only supports 1:1 physical-virtual mappings for boot
modules and can't cope with modules lying above 3G physical. There's no
way to map them because of the 1:1 enforcement. Set max physical relocation
memory to 2G.

(2)
Support "256M" commandline option for the upper physical address. The seL4
kernel fails to load if the modules are at the end of the physical memory.
So add a 256M option as upper bound, so that the kernel behind finds free
memory (if you have a machine with less then 256M you're lost of course,
for various other reasons).